Resolved multiple warnings

+++ model/model-core/src/main/java/com/castlemock/model/core/utility/parser/expression/RandomDoubleExpression.java
@@ -54,7 +54,7 @@ public class RandomDoubleExpression extends AbstractExpression {
         }
 
         final int randomValue = RANDOM.nextInt(maxLength) + minLength;
-        return Double.toString((double) randomValue);
+        return Double.toString(randomValue);
     }
 
     /**
